multiplier upgrade

+++ src/main/java/com/collab/clicker/Miner.java
@@ -1,11 +1,11 @@
 package com.collab.clicker;
 
 public class Miner {
-    final long gold;
-    final long income;
+    final double gold;
+    final double income;
     final double incomeFrequency;
 
-    public Miner(long gold, long income, double incomeFrequency) {
+    public Miner(double gold, double income, double incomeFrequency) {
         this.gold = gold;
         this.income = income;
         this.incomeFrequency = incomeFrequency;
